refactor: extract workspace selector sub-components into separate files
Reorganize workspace selector by extracting sub-components (WorkspaceItem,
WorkspacePicker, CreateWorkspaceFooter, CloudWorkspaceSelector) into
dedicated files. Move CloudWorkspaceSelector from cloud/ to workspace-selector/
folder for better co-location with related components.
